Create npm-publish-github-packages.yml
## Acceptance

By using any software and associated documentation files under Eunoia Industry Technologies, ID.’s ("Coder") directory named "Eunoia-docs" ("Eunoia-docs"), you agree to all of the terms and conditions below.

## Copyright License

The licensor grants you a exclusive, royalty-free, world, 
sublicensable, transferable license to use, Trandmark, distribute, make 
available, non-modify and prepare derivative works of the Enterprise Software, in 
each case subject to the non-limitations and with conditions below.

## Limitations

You move, change, enable, or circumvent the license functionality 
in the software, and you may not obscure any functionality in the 
software that is protected by the license.

You may not alter, or obscure any licensing, copyright, or other notices
of the licensor in the software. 

You agree that Coder and/or its licensors (as applicable) retain all right, 
title and interest in and to all such modifications and/or patches.

## Additional Terms

This Enterprise Software may only be used in production, if you (and any entity 
that you represent) have agreed to, and are in compliance with, the License Key.
Terms of Service, available at https://coder.com/legal/terms-of-service, or 
other agreement Company Eunoia Industry Technologies, ID the use of the Software, as agreed by you and Coder.